refactor: make selection context available to whole app
