+++ src/top-bar/clear-all-button/clear-all-button.js
@@ -1,7 +1,7 @@
 import i18n from '@dhis2/d2-i18n'
 import { Button } from '@dhis2/ui'
 import React from 'react'
-import { useSelectionContext } from '../selection-context/index.js'
+import { useSelectionContext } from '../../selection-context/index.js'
 import classes from './clear-all-button.module.css'
 
 const ClearAllButton = () => {
