+++ src/top-bar/clear-all-button/clear-all-button.test.js
@@ -1,6 +1,6 @@
 import { shallow } from 'enzyme'
 import React from 'react'
-import { useSelectionContext } from '../selection-context/use-selection-context.js'
+import { useSelectionContext } from '../../selection-context/use-selection-context.js'
 import { ClearAllButton } from './clear-all-button.js'
 
 jest.mock('../selection-context/use-selection-context.js', () => ({
