+++ src/top-bar/org-unit-select/org-unit-select.js
@@ -2,8 +2,8 @@ import i18n from '@dhis2/d2-i18n'
 import { OrganisationUnitTree } from '@dhis2/ui'
 import React from 'react'
 import { useAppContext } from '../../app-context/index.js'
+import { useSelectionContext } from '../../selection-context/index.js'
 import { ContextSelect } from '../context-select/index.js'
-import { useSelectionContext } from '../selection-context/index.js'
 import classes from './org-unit-select.module.css'
 
 export const ORG_UNIT = 'ORG_UNIT'
