+++ src/top-bar/workflow-select/workflow-select.js
@@ -2,8 +2,8 @@ import i18n from '@dhis2/d2-i18n'
 import { Menu } from '@dhis2/ui'
 import React from 'react'
 import { useAppContext } from '../../app-context/index.js'
+import { useSelectionContext } from '../../selection-context/index.js'
 import { ContextSelect } from '../context-select/index.js'
-import { useSelectionContext } from '../selection-context/index.js'
 import { WorkflowSelectOption } from './workflow-select-option.js'
 import classes from './workflow-select.module.css'
