+++ src/top-bar/workflow-select/workflow-select.test.js
@@ -3,8 +3,8 @@ import { shallow } from 'enzyme'
 import React from 'react'
 import { useAppContext } from '../../app-context/index.js'
 import { readQueryParams } from '../../navigation/read-query-params.js'
+import { useSelectionContext } from '../../selection-context/index.js'
 import { ContextSelect } from '../context-select/context-select.js'
-import { useSelectionContext } from '../selection-context/index.js'
 import { WorkflowSelectOption } from './workflow-select-option.js'
 import { WORKFLOW, WorkflowSelect } from './workflow-select.js'
